major refactor - all working again
