+++ bots/discord_alerts.py
@@ -4,11 +4,14 @@ import discord
 from discord.ext import tasks
 from datetime import time as dtime
 from zoneinfo import ZoneInfo
+
+# -------- Load config --------
 load_dotenv()  # loads .env if present
 
 TOKEN = os.getenv("DISCORD_ALERTS_TOKEN")
 CHANNEL_ID = int(os.getenv("DISCORD_ALERTS_CHANNEL_ID", "0"))
 
+# -------- Discord client --------
 intents = discord.Intents.default()
 client = discord.Client(intents=intents)
 
@@ -16,6 +19,8 @@ client = discord.Client(intents=intents)
 EUL_TZ = ZoneInfo("Europe/London")
 EIGHT_AM_UK = dtime(hour=8, minute=0, tzinfo=EUL_TZ)
 
+
+# -------- Scheduled morning message --------
 @tasks.loop(time=EIGHT_AM_UK)
 async def morning_alert():
     if CHANNEL_ID == 0:
@@ -24,15 +29,21 @@ async def morning_alert():
     if channel:
         await channel.send("🌤️ Good morning! Time for your daily check in:")
 
+
+# -------- Bot ready event --------
 @client.event
 async def on_ready():
     print(f"✅ Logged in as {client.user}")
     channel = client.get_channel(CHANNEL_ID)
-    await channel.send("🐾 CatPanel bot is online.")
+    if channel:
+        await channel.send("🐾 CatPanel bot is online.")
+
     check_alerts.start()
     if not morning_alert.is_running():
         morning_alert.start()
 
+
+# -------- Check and relay alerts --------
 @tasks.loop(seconds=30)
 async def check_alerts():
     if CHANNEL_ID == 0:
@@ -40,6 +51,7 @@ async def check_alerts():
     try:
         with open("data/alerts.log", "r") as f:
             lines = [ln.strip() for ln in f if ln.strip()]
+
         if not lines:
             return
 
@@ -50,14 +62,17 @@ async def check_alerts():
                 time_str = parts[1]
                 ts = " ".join(parts[3:5]) if "at" in parts else ""
                 msg = f"🐾 **Fed confirmed** — {time_str} ✅  ({ts})"
+
             elif line.startswith("UNSET"):
                 parts = line.split()
                 time_str = parts[1]
                 ts = " ".join(parts[3:5]) if "at" in parts else ""
                 msg = f"⚠️ Feed toggle **cleared** — {time_str}  ({ts})"
+
             elif line.startswith("Missed feeding for"):
                 # old format
                 msg = f"🚨 {line}"
+
             else:
                 msg = line  # fallback
 
@@ -65,7 +80,11 @@ async def check_alerts():
 
         # clear file after sending
         open("data/alerts.log", "w").close()
+
     except FileNotFoundError:
+        # alerts.log might not exist yet — ignore
         pass
 
+
+# -------- Run --------
 client.run(TOKEN)
