+++ templates/index.html
@@ -106,7 +106,7 @@
     </section>
 
   </main>
-  <script src="/static/main.js"></script>
+  <script type="module" src="/static/scripts/main.js"></script>
 </body>
 
 </html>
